Split JS files following Jest update
Wasn’t possible before for… reasons. ¯\_(ツ)_/¯
