+++ .eslintrc.json
@@ -7,6 +7,7 @@
     "extends": "eslint:recommended",
     "parserOptions": {
         "sourceType": "module",
+        "ecmaVersion": 12,
         "ecmaFeatures": {
             "impliedStrict": true
         }
@@ -15,7 +16,7 @@
     "rules": {
         "indent": [2, 2, { "SwitchCase": 1 }],
         "linebreak-style": [2, "unix"],
-        "quotes": [2, "single"],
+        "quotes": [2, "single", { "allowTemplateLiterals": true }],
         "semi": [2, "never"],
 
         "array-bracket-spacing": [2, "never"],
